Fix regression when conference has a password or auth
This partially reverts eec6a270c5b2c7c93891dee1f9b49cfd351f7cff.

The problem is that when a conference is protected by a password or has
auth enabled, the "joined" event will come later, when the user
authenticates, so the 10s timeout will kick in.

This is not ideal because we react to the onLoad event of an iframe,
which doesn't tell us much of *what* is actually loaded, but a new event
would be required to properly fix this.