refactor: move all dependencies that are webpacked into dev-deps
This reduces the shipped binary by 56MB (Windows, Linux) and 2x56MB on mac

Previously all the dependencies were webpacked and minified in the asar in the
build folder and in its full installed version unused under node_modules.

Only keep the dependencies that are externally required in the node-modules
folder.

Signed-off-by: Christoph Settgast <csett86@web.de>