refactor: reduce shipped artifact size
This moves all webpacked deps to be automatically included from the asar
and thus reduces the shipped binary by 14MB on Windows, 16 MB on Mac and
12MB on Linux. The installed size on mac is reduced by 105 MB.

Previously all the dependencies were webpacked and minified in the asar in the
build folder and in its full installed version unused under node_modules.

Only keep the dependencies that are externally required in the node-modules
folder.

Signed-off-by: Christoph Settgast <csett86@web.de>