mac: Enable autoupdate by sign and notarize via github action
Signed and notarized binaries are the precondition for autoupdates on
mac. Additionally Gatekeeper on 10.15+ is happy and allows to open the
app instead of blocking it.

The notarize step is added unconditionally, as it only emits a warning if
the notarization API key is not set, but it does not break the build.

This is an upstreaming of https://github.com/csett86/jitsi-meet-electron
where it worked since March 2020.

The required github secrets (signing key, cert and notarize API key) are
best described at https://github.com/samuelmeuli/action-electron-builder#code-signing

On CI, only sign if not triggered by pull request, as these will fail (as secrets
are not available to pull request builds).