feat: force prejoin page instead of profile and av settings
The settings and the prejoin page which is forced on many instances
including meet.jit.si did conflict, as you could set values in the
settings, then try to override them in den prejoin page only to notice
that it would not work. Clean this up by only having the prejoin page
where all settings can be done before joining the conference.

Closes: #687