chore(deps) update electron to 25.3.0
Contains the fix for the hard crash on wayland when starting screenshare. Unfortunately it now hard-crashes when stopping screensharing.

The hard crash on starting screensharing is fixed, but its still relatively unreliable to share the screen under wayland. Background is electron/electron#39043, and same as described there two Pipewire pickers open, and the user has to select and accept both, then be quick in the jitsi selector as well and then it works. If one Pipewire picker is cancelled by the user, the process hangs. If the user is too slow in the jitsi selector, the two Pipewire pickers pop up again.

See: #829