chore(deps) update electron to 27.0.3
Contains electron 26 and 27 updates, for details see

https://www.electronjs.org/blog/electron-26-0
https://www.electronjs.org/blog/electron-27-0

Contains a fix for Linux that lead to crashes when the graphics drivers changed
https://github.com/electron/electron/pull/40467

Mainly this should help with further bugfixes in webrtc as the contained
Chromium is implicitly upgraded from 114 to 118.

In wayland / pipewire terms we have only minor additions, eg:

WaylandWindowDecorations by default: https://github.com/electron/electron/pull/39582

which means we can remove the enable flag we had on this in main.js as its now
default.